Interface children, childX, and childY integer arrays have been converted to LinkedLists for the ease of adding, removing, rearranging, and editing children.
The interface cache has been changed from an RSInterface array to a
HashMap for the ease of adding and removing interfaces from the
interface cache.

Started working on creating new interface types (rs2.rsinterface.type
contains classes for creating new interface types). TextType is the only
one right now and it is untested.

Importing a .rsi file will open a file chooser for you to make importing
much quicker and easier.

I've added a "Cache" menu to the menubar. As of right now you can
recompile,
import, and export new interface.jag archives and new media.jag
archives. When exported they are sent to ./export/. When importing a
file chooser pops up to choose the file.